+++ actionmailer/lib/action_mailer/railtie.rb
@@ -18,9 +18,9 @@ module ActionMailer
       options.assets_dir      ||= paths["public"].first
       options.javascripts_dir ||= paths["public/javascripts"].first
       options.stylesheets_dir ||= paths["public/stylesheets"].first
-      options.preview_enabled = Rails.env.development? if options.preview_enabled.nil?
+      options.show_previews = Rails.env.development? if options.show_previews.nil?
 
-      if options.preview_enabled
+      if options.show_previews
         options.preview_path  ||= defined?(Rails.root) ? "#{Rails.root}/test/mailers/previews" : nil
       end
 
@@ -39,7 +39,7 @@ module ActionMailer
 
         options.each { |k,v| send("#{k}=", v) }
 
-        if options.preview_enabled
+        if options.show_previews
           app.routes.append do
             get '/rails/mailers'         => "rails/mailers#index"
             get '/rails/mailers/*path'   => "rails/mailers#preview"